rename: ProfileImage 엔티티를 member 패키지로 이동
