+++ packy-api/src/main/java/com/dilly/auth/application/AuthService.java
@@ -18,7 +18,6 @@ import com.dilly.auth.dto.response.SignInResponse;
 import com.dilly.auth.model.AppleAccountInfo;
 import com.dilly.auth.model.AppleToken;
 import com.dilly.auth.model.KakaoResource;
-import com.dilly.gift.ProfileImage;
 import com.dilly.global.exception.UnsupportedException;
 import com.dilly.global.response.ErrorCode;
 import com.dilly.global.utils.SecurityUtil;
@@ -28,6 +27,7 @@ import com.dilly.jwt.domain.JwtReader;
 import com.dilly.jwt.domain.JwtWriter;
 import com.dilly.jwt.dto.JwtResponse;
 import com.dilly.member.Member;
+import com.dilly.member.ProfileImage;
 import com.dilly.member.Provider;
 import com.dilly.member.Status;
 import com.dilly.member.domain.MemberReader;
