+++ packy-api/src/main/java/com/dilly/auth/dto/request/SignupRequest.java
@@ -1,7 +1,7 @@
 package com.dilly.auth.dto.request;
 
-import com.dilly.gift.ProfileImage;
 import com.dilly.member.Member;
+import com.dilly.member.ProfileImage;
 import com.dilly.member.Provider;
 
 import io.swagger.v3.oas.annotations.media.Schema;
