+++ packy-domain/src/main/java/com/dilly/gift/dao/ProfileImageRepository.java
@@ -2,7 +2,7 @@ package com.dilly.gift.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.dilly.gift.ProfileImage;
+import com.dilly.member.ProfileImage;
 
 public interface ProfileImageRepository extends JpaRepository<ProfileImage, Long> {
 }
